# Draw pen lines via fragment shader