--- a/src/shaders/sprite.vert
+++ b/src/shaders/sprite.vert
@@ -1,22 +1,26 @@
-uniform mat4 u_projectionMatrix;
 uniform mat4 u_modelMatrix;
+
+#ifdef DRAW_MODE_line
+uniform vec2 u_stageSize;
+#endif
+
+#ifndef DRAW_MODE_line
+uniform mat4 u_projectionMatrix;
+#endif
 
 attribute vec2 a_position;
 attribute vec2 a_texCoord;
 
 varying vec2 v_texCoord;
 
-#ifdef DRAW_MODE_lineSample
-uniform float u_positionScalar;
-#endif
-
 void main() {
-    #ifdef DRAW_MODE_lineSample
-    vec2 position = a_position;
-    position.y = clamp(position.y * u_positionScalar, -0.5, 0.5);
-    gl_Position = u_projectionMatrix * u_modelMatrix * vec4(position, 0, 1);
+    #ifdef DRAW_MODE_line
+    vec4 screenCoord = u_modelMatrix * vec4(a_position, 0, 1);
+
+    gl_Position = screenCoord;
+    v_texCoord = ((screenCoord.xy * 0.5) + 0.5) * u_stageSize;
     #else
     gl_Position = u_projectionMatrix * u_modelMatrix * vec4(a_position, 0, 1);
-    #endif
     v_texCoord = a_texCoord;
+    #endif
 }