### Resolves

Kinda resolves #420?
Also resolves https://github.com/LLK/scratch-gui/issues/4313 [EDIT: and now resolves #306]

### Proposed Changes

This changes the PenSkin's line-drawing routine from a primarily vertex-based operation to a fragment-based one. Lines are now drawn via a screen-space pixel shader that provides proper antialiasing.

```PenSkin.drawLineOnBuffer()``` now draws a single ~~axis-aligned~~ quad. The quad's fragment shader, given the endpoints, color, and thickness of the line, properly computes antialiased line coverage for each pixel inside it.

This has two effects:
1. It gets rid of all the jaggies on the edges of each line
2. It makes dots truly circular

This still doesn't completely match 2.0. In 2.0, "size 1" lines appear thicker at certain angles, and pen *points* render at a different offset from pen *lines*. The former might be worth investigating, but in my opinion the latter behavior is too confusing/annoying to be worth emulating.

On My Machine™, it runs [236783324](https://scratch.mit.edu/projects/236783324/) at the same speed, suggesting the bottleneck is in the VM now, but more comprehensive benchmarking across different GPUs would be wise.

~~I've done very little WebGL, so a lot of cleanup/refinement is probably necessary.~~

Current | This PR | Scratch 2.0
---|---|---
![1-before](https://user-images.githubusercontent.com/25993062/56542622-fdae1c00-653c-11e9-8171-728f17ae9f82.png) | ![1-after](https://user-images.githubusercontent.com/25993062/56499639-4205d180-64d5-11e9-8a8d-f9b768ae75e2.png) | ![1-scratch2](https://user-images.githubusercontent.com/25993062/56455573-ea793180-632d-11e9-9b41-cf48f651ba1a.png)
[Pen stroke test](https://scratch.mit.edu/projects/303669200/)
![2-before](https://user-images.githubusercontent.com/25993062/56455592-2ad8af80-632e-11e9-8e00-deb57e42af30.png) | ![2-after](https://user-images.githubusercontent.com/25993062/56499645-47631c00-64d5-11e9-9ffe-0c29abdbdcbb.png) | ![2-scratch2](https://user-images.githubusercontent.com/25993062/56455595-31672700-632e-11e9-8f18-777277108555.png)
[Pen stroke test 2](https://scratch.mit.edu/projects/303711585/)
![3-before](https://user-images.githubusercontent.com/25993062/56455613-8145ee00-632e-11e9-96a1-927ed8eb1b9a.png) | ![3-after](https://user-images.githubusercontent.com/25993062/56499658-4d58fd00-64d5-11e9-9db6-4f53f9d964a0.png) | ![3-scratch2](https://user-images.githubusercontent.com/25993062/56455616-85720b80-632e-11e9-8980-47fce8bb97ed.png)
[Pen roundness test](https://scratch.mit.edu/projects/303711508/)

EDIT: [This](https://scratch.mit.edu/projects/55619918/) may be a good test project. Make sure to run this multiple times-- there's a lot of weird variance. It looks like most of the time (in both the current code and this PR's) is spent in setUniforms, but fixing that would likely require UBOs, which are a WebGL 2 feature.