@adroitwhiz 2019-12-01
It looks like disabling antialiasing (#502) also had an effect on pen lines:
![image](https://user-images.githubusercontent.com/25993062/69911331-d7f8c780-13e7-11ea-9038-9b1677217094.png)