@fsih 2020-02-04
@BryceLTaylor We're considering putting in this PR that should improve the appearance of pen lines, but we need to benchmark it first to make sure that it doesn't significantly impact pen performance, especially on devices with lower-power GPUs. Can we work together with you sometime to design benchmarking tests?

@cwillisf @fsih