@adroitwhiz 2020-02-05
Also, are you still interested in `PenSkin` code cleanups (e.g. #450)? There's a lot of unused code (like `drawStamp`) and unnecessary code (like the rectangle-drawing stuff) that could be removed.