@adroitwhiz 2020-02-11
I've done some benchmarking of my own to compare the performance of:
- The old line implementation
- The new line implementation, with the bounds calculation done in JS + uploaded to the vertex shader via a 4x4 matrix uniform
- The new line implementation, with the bounds calculation done in the vertex shader

## "Workstation" (AMD Ryzen 7 1700 + NVIDIA GTX 780 Ti)

Run # | Firefox (old lines) | Firefox (new lines; bounds in JS) | Firefox (new lines; bounds in VS)
--- | --- | --- | ---
1 | 44737 | 43776 | 49984
2 | 51013 | 51552 | 52207
3 | 51153 | 49608 | 60704
4 | 57501 | 52822 | 55869
5 | 54206 | 52198 | 56861
avg. | 51722 | 49991 | 55125

Run # | Chrome (old lines) | Chrome (new lines; bounds in JS) | Chrome (new lines; bounds in VS)
--- | --- | --- | ---
1 | 105232 | 82110 | 99655
2 | 104882 | 92315 | 96482
3 | 102243 | 85058 | 93570
4 | 100148 | 88413 | 101246
5 | 105377 | 90816 | 93938
avg. | 103576 | 87742 | 96978

Run # | Chrome SwiftShader (old lines) | Chrome SwiftShader (new lines; bounds in JS) | Chrome SwiftShader (new lines; bounds in VS)
--- | --- | --- | ---
1 | 13845 | 15255 | 13879
2 | 13930 | 15173 | 14053
3 | 13979 | 14622 | 14083
4 | 13883 | 15171 | 14062
5 | 13988 | 15242 | 14243
avg. | 13925 | 15093 | 14064

Browser | Avg. speedup over old lines (bounds in JS) | Avg. speedup over old lines (bounds in VS)
--- | --- | ---
Firefox | -3.3% | 6.6%
Chrome | -15.3% | -6.4%
Chrome SwiftShader | 8.4% | 1.0%

## Old laptop (Intel Core 2 Duo T5800 + Intel Express 965)

Note that SwiftShader was enabled by default in Chrome due to GPU blacklisting. Firefox

Run # | Firefox (old lines) | Firefox (new lines; bounds in JS) | Firefox (new lines; bounds in VS)
--- | --- | --- | ---
1 | 4730 | 3794 | 5152
2 | 4359 | 3437 | 6106
3 | 5211 | 4530 | 3747
4 | 5158 | 4619 | 3618
5 | 5380 | 3833 | 6065
avg. | 4968 | 4043 | 4938

Run # | Chrome (old lines) | Chrome (new lines; bounds in JS) | Chrome (new lines; bounds in VS)
--- | --- | --- | ---
1 | 21631 | 19560 | 21992
2 | 21960 | 22909 | 23525
3 | 22261 | 20026 | 23983
4 | 22721 | 24325 | 24186
5 | 23059 | 24182 | 24017
avg. | 22326 | 22200 | 23541

Run # | Chrome SwiftShader (old lines) | Chrome SwiftShader (new lines; bounds in JS) | Chrome SwiftShader (new lines; bounds in VS)
--- | --- | --- | ---
1 | 4068 | 3789 | 3558
2 | 6423 | 5372 | 4224
3 | 5570 | 6439 | 4371
4 | 5861 | 5905 | 4397
5 | 6961 | 6439 | 4433
avg. | 5777 | 5589 | 4197

Browser | Avg. speedup over old lines (bounds in JS) | Avg. speedup over old lines (bounds in VS)
--- | --- | ---
Firefox | -18.6% | -0.6%
Chrome | -0.6% | -5.4%
Chrome SwiftShader | -3.3% | -27.3%

In general, it appears that:
- The new pen line implementation is very slightly slower than the old one, but only by a few percent.
- Moving the pen line bounds calculation to the vertex shader speeds up Firefox, but slows down SwiftShader (Chrome's software renderer). It seems to have mixed effects on Chrome + hardware rendering depending on the hardware.