@cwillisf 2020-02-20
> Also, are you still interested in `PenSkin` code cleanups (e.g. #450)? There's a lot of unused code (like `drawStamp`) and unnecessary code (like the rectangle-drawing stuff) that could be removed.

Sorry -- missed this comment. We're sort of re-evaluating our process for priorities so it's hard to know what should or shouldn't be prioritized until the dust settles. That said, I think code cleanups, such as addressing #450, are welcome but won't necessarily get a high priority. As long as you're OK with that, go for it ;)